Move theme to a specific file
